Implemented tests on new version of CSSRule
